Timeout, delay and 3 reruns
Timeout and delay should now be implemented for the cluster. Have not had enough time to see if it works properly. The logic should be there though. The image is not exited but there is a condition for where it should.

NOTE: Had to revert the previous version of this commit so it became a bit larger than expected. Now the rerun X times is in a dev-mode where after the tests pass a counter goes up, when the counter is larger than 3 it prints a message.